requirements.txt
google_api_python_client==2.172.0
html2text==2025.4.15
pandas==2.3.0
python-dotenv==1.1.0
pysafebrowsing